BasisImporter: Y-flip RGBA8, BC1, BC3, BC4 and BC5 on load.
Compared to DdsImporter this has the additional complication of being
able to pick a different format for every image import so the warnings
are done all inside image*D(). The Basis and KTX containers also can
contain orientation metadata, which (together with that well-known basis
bug where it doesn't properly save the metadata to KTX files) adds some
more axes to the testing data.

On the other hand, as Basis doesn't support 3D textures (even though it
claims so, you see a pattern here, right?), there isn't any Z-flip
necessary, so for that part it only tests that the right dimension gets
flipped for 3D data, and not any other.

As with DdsImporter, this flipping can be turned off either by the
assumeYUp option (or by having Y up metadata in the file).