Ktx{Importer,ImageConverter}: support 3D ASTC formats.
I need them to test Z-flip of volume textures and the state of the
ecosystem is just great -- ARM's astcenc can generate files with 3D ASTC
blocks, but doesn't support *.ktx2 output, only *.astc. And then,
KTX-Software doesn't support *.astc input. So the only way to get those
files is to just implement the support in my own tools.

I asked about the Vulkan enums for 3D ASTC back in 2019 and got a link
to a private fork of a Vulkan header containing them. It's 2023 now and
those enums are still only in that private fork. Exquisite.