KtxImporter: Y-flip BC1/2/3/4/5 blocks on load.
Basically a combination of the changes done in DdsImporter and
BasisImporter, together with extra logic that prints a warning in case
X-axis flip is desired (for which, so far, there aren't any APIs). Since
KTX2 supports 3D ASTC formats as well, compared to DdsImporter it also
prints a warning in case a Z-flip of a volume format would be needed
(which isn't implemented yet either, not even Y-flip is implemented for
ASTC yet).

As with the two others, the flips (or warnings) can be configured with
the assumeOrientation option.