BasisImporter: print a warning if KTX2 orientation metadata are missing.
As the comment says -- ideally this would just pick a default value for
the missing metadata and go on. But because the *main implementation* of
the Basis format doesn't care about properly writing the metadata in,
it's not clear at all what the orientation could be. Thus, printing a
warning is a reasonable thing to do, because if the file looks weird,
the warning may suggest why it does.