JpegImageConverter: test also the actual output file contents.
In order to avoid bugs similar to Acropalypse, it's not enough to just
check that the file imports to given pixel data again. Plus it prepares
for upcoming metadata support (colorspaces, orientation...).

This however results in slight differences in the output files across
different libjpeg versions, hopefully there won't be too many of them.