PngImageConverter: test all output file contents.
Similar to the JpegImageConverter change in the previous commit, and a
followup to 2291aa91600ce32ebc3f229cb2c440c13aaeb5fd where I was just
lazy. This makes the tests more robust so I can do changes in the plugin
internals, and prepares also for upcoming metadata support (e.g.,
marking the files as sRGB).