GltfImporter: rethink handling of attributes disallowed in morph targets.
The import now fails instead of turning them into custom attributes
(that then wouldn't have a matching base attribute anymore) and printing
a confusing error message that implies their format is wrong while the
attribute itself is disallowed. The message is now clearly stating that
given attribute is not allowed to be within a morph target.

There's also still the case where the attribute is allowed in a morph
target but has an unsupported format. That's still handled the same way,
turning it into a custom attribute.

As joints/weights aren't allowed in morph targets by glTF spec itself,
this shouldn't be a problem in practice. And the object ID attribute
also isn't something directly supported by glTF, so morphing it is also
rather unlikely.

If this ever actually becomes a problem, we can reconsider and
actually make the attributes imported as custom. This is exactly what
happened with invalid formats before, originally they caused an import
error but because there were assets produced by broken tools containing
non-normalized 16bit colors, I changed the behavior to import them as
custom.