AssimpImporter: a better fix for try_compile() on Windows.
Instead of setting CMAKE_BUILD_TYPE which was done in
8e5cf7ac9820be5cf1f8728db83a284eb9120090 and which didn't really work on
multi-config builds unless a matching Assimp library was present for
every build type, try_compile() is now making a static library instead
of an executable, circumventing all the _ITERATOR_DEBUG_LEVEL checks
that are done at link time.

It requires CMake 3.6, but that's a non-issue on Windows as there people
commonly use a lot more recent releases.