{Free,StbTrue}TypeFont: use Utility::copy() for copying glyph data.
Safer, less error prone and definitely not slower than the manual for
loop. Not to mention it's trivial to extend for color data such as
emoji.

In case of FreeTypeFont there originally were assertions for glyph
bitmap sizes being "not too far". Can't seem to trigger any size
mismatch with the new code however, will investigate once it happens.