HarfBuzzFont: add a test TODO.
I have a vague feeling that the offsets had a sign flipped or some such
when I tried that feature with a Zalgo text back in 2012. Or maybe that
was *the other* codebase, and here it was correct?

Won't really know until the API is capable of filling a glyph cache with
glyph IDs, to include various combining marks for a visual test.