Png{Importer,ImageConverter}: improve test files for alpha formats.
Having alpha either FF or 00 (and the pixels all zero in that case as
well) doesn't really help with testing anything non-trivial, FFS.

Also updating tests that relied on these broken files. Some of them
still use the originals, just renamed, some are updated to the files
with non-trivial alpha.