PngImporter: add an option to premultiply alpha on import.
Of course such a seemingly trivial task involved a day of investigation
in the libpng documentation. Yes, there's the additional complication
with sRGB-encoded values, but that I expected. What I didn't expect was
having to fake some "screen gamma" in order for the library to give me
back unchanged pixel values if there's no alpha channel or if there's
just opaque pixels.