StanfordImporter: delegate ASCII files to Assimp.
My evil master plan to convince people to stop using ASCII PLYs by
making it really annoying to open them didn't work out. So, instead of
suffering through implementing a parser, do the next easiest thing and
delegate to Assimp.

It'll be awfully slow, it'll crash on bad input, etc., but that's what
you get for deliberately using a bad variant of an otherwise great
format. I'm not to blame.