PngImporter: add support for alphaMode=premultipliedLinear.
The thing is that setting alphaMode to premultiplied can result in the
pixel value going from (255, 255, 255, 128) to (186, 186, 186, 128).
Which is totally fine if the output is interpreted as sRGB, but causes
strange bright halos when it's not.

Ideally Magnum would be sRGB-aware everywhere, all existing end-user
code would be properly sRGB-aware as well, *and* all editors would
correctly export the gAMA chunk to distinguish linear and sRGB assets.
(Inkscape, looking at you.) But the reality is different, so this is
needed.