OpenExrImageConverter: adapt test files for libdeflate 1.19.
OpenEXR 3.2 switched from zlib to libdeflate, so I updated the test
files. But shortly after (3.2.1) it started failing again, and I just
couldn't see anything that would become different in that version. Turns
out it's libdeflate 1.19 producing a different output than 1.18. And
libdeflate version is impossible to check for, so I'm just updating to
latest with a hope that systems that have OpenEXR 3.2+ also have
libdeflate 1.19+.