PngImporter: require building with libpng 1.5.4+.
Or with a version of libpng that doesn't have the functionality we need
compiled out -- libpng 1.5.4 is from 2011. Which is 99% of use cases, so
I don't think it makes sense to try to gracefully fail for those, but
instead of failing due to random identifiers missing, provide clear
error messages about what might be wrong.

Like, certain software packages still rely on libpng 1.2 (such as the
Steam runtime), so it's not completely unlikely that the plugin might
get built against that version by accident.