BasisImageConverter: disable asserts in Basis' OpenCL code
When it doesn't find a GPU device, it falls back to a CPU device and
prints a nice error along the way. Unfortunately, all OpenCL-related
errors are accompanied by an assert that can't be turned off. There's
BASISU_OPENCL_ASSERT_ON_ANY_ERRORS but that's always defined to 1 at the
top of the file...

So disable asserts via NDEBUG to actually get to use (and test) the
OpenCL code in debug builds.