BasisImageConverter: fix a non-deprecated build.
The enum was made private in bdc2bfd7731884aff757e36ec6d106abe7e8ac8c and
unfortunately GCC 4.8 (which is the only compiler that tests a
non-deprecated build on the CI) happily accepted the enum being used
outside of the class. All other compilers treat that as an error.