modules: populate library DLL locations for dynamic builds on Windows.
This involves checking if MAGNUM_*_BUILD_STATIC is set and deciding
based on that. Since the same had to be done for static plugin loading
already, that logic is now in the same place for both libraries and
plugins. As with core Magnum itself, plugins don't have this treatment
so far as they are a pure runtime dependency.