GltfImporter: expand tests for duplicate property cleanup with escapes.
The current code works, but I want to get rid of std::stable_sort() and
need to ensure I'm not comparing what I shouldn't compare, such as
string view data pointers where one of them might be pointing to the
input JSON but the other not.

The original mesh-duplicate-attributes.gltf file was copied to
(deprecated) TinyGltfImporter tests along with a *.bin, since it can no
longer be shared between the two.