GltfImporter: only copy actually used vertex data ranges.
The original way, where it simply copied the whole range, was done to
optimistically prepare for a case where the *.bin would be memory-mapped
and the data import would be just passing the view slice through.

But reality is different, and with for example the Buggy.gltf, which
contains over a hundred meshes, the total vertex data size was over 800
MB (!) while the model is just 7 MB. I tried to find some less
complicated solution such as just remembering which buffer views get
used and then copying those, but that didn't work either (with this
Buggy.gltf model in particular) and had to really do it the hard way, by
finding which data ranges overlap, which not, and then copying those
slices one after another to the output array.

Co-authored-by: Hugo Amiard <hugo.amiard@wonderlandengine.com>