AssimpImporter: provide the UsdImporter alias.
The AnySceneImporter now directs USD files to it, and it's supported
since Assimp 5.4.3. Unfortunately the package in Arch doesn't have it
enabled (likely because it's off by default, because the "tinyusd" code
takes AGES to compile, and because some lunatic enabled -Werror there by
default, which OF COURSE makes it fail to compile).