HarfBuzzFont: add a dedicated test producing X and Y glyph offsets.
First I tried to reduce a "Zalgo" text to a reasonable size but then
thought "what if I just add the diacritics two more times", and turns
out that's exactly what Zalgo does. So, yeah.

This investigation also concluded the TODOs in FreeTypeFont and
StbTrueTypeFont -- they have no such thing as glyph offsets, and what I
*thought* was the offset, turns into bitmap offset and is stored in the
glyph cache already (such as, in this case, the X and Y shift of the
diacritics when it gets actually rendered).