ResvgImporter: new, Rust-based plugin for importing SVGs.
I expected a lot from this library, being it a Rust library. In reality
the disappointed-comment-to-lines-of-code ratio is one of the worst
among all plugins. Sigh.

Furthermore, this plugin is actually not tested at all on any of the
CIs. On Ubuntu it's available since 22.04 but they don't build the C
interface so it's like if it wasn't there at all. It's in Homebrew, but
installing it requires making a `brew update`, and then it proceeds with
building LLVM from scratch. Fuck that. And the official releases are
just a single self-contained executable, no library or C header either.
So this is so far just tested on my Arch, which makes it an immediate
tech debt. Yay.