package: enable ResvgImporter in dev Arch PKGBUILDs and in Homebrew.
Being it a Rust thing, it's apparently either not included in repos at
all or included without the C interface. And if it's available the extra
dependencies make it potentially very annoying to get, so it's only
enabled in the dev PKGBUILDs and not in the AUR package. For Homebrew
it's opt-in because when I tried to install it on my CI, it attempted to
build LLVM from scratch. Yay.