GltfImporter: adapt to Utility::Json API robustness changes.
Note that there's a breaking change in the importerState() API -- it's
now a pointer to Utility::JsonTokenData, not JsonToken, so you need to
update your code. Unfortunately due to the void* typelessness it doesn't
get caught at compile time, so be sure to grep for all importerState()
and SceneField::ImporterState use and update accordingly.