Faad2Importer: what a convoluted way to append to an array.
The original code before 12b9db7bb5760acd49ba641cbbc544a8404ce635 was
doing a std::vector::resize() followed by std::copy_n() which isn't
really great either, but that's understandable because having to form a
begin and end iterator to use std::vector::insert() neither of which is
really a pinnacle of API usability is bad enough as it is. And this was
just an unfortunate result of porting away from that mess.

As I'm saying, often it's better to just start from scratch instead of
trying to port STL-heavy code to something sane.