Don't use the deprecated DefaultInit initialization.
BasisImporter used it instead of NoInit for historical reasons, as
NoInit used to cause a custom deleter to be used even for trivially
constructible types.

One case where it was used in UfbxImporter is now (implicit) NoInit as
well, as the arrayShrink() is called on an array of a trivial type.

All other cases are with class types, and thus DefaultInit was
equivalent to ValueInit.