KtxImageConverter: this error path is no longer triggerable.
As the now-deleted comment said -- the error path was only executed by
passing an invalid CompressedPixelFormat. But now, as CompressedImage*
requires the CompressedPixelFormat to be valid in order to fetch block
properties for it, KtxImageConverter can only get an image that has a
valid format, and thus none of this is possible to hit.